desktop_app/requirements.txt
PySide6>=6.6
pyqtgraph>=0.13
pyserial>=3.5
pygame>=2.5
